remove unnecessary csv import folder, connect import csv w/ endpoints, format css, update path to csv import in team overview
