--- CodebustersPAS-ClientApp/src/pages/Teacher/CreateTeam/CSV_Import/src/App.jsx
@@ -1,15 +1,0 @@
-// src/App.jsx
-
-import React from 'react';
-import CSVUpload from './Components/CSVUploads';
-
-const App = () => {
-  return (
-    <div className="App">
-      <h1>Team Management</h1>
-      <CSVUpload />
-    </div>
-  );
-};
-
-export default App;
