+++ CodebustersPAS-ClientApp/src/pages/Teacher/CreateTeam/TeamOverview.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import CSVUpload from './CSVUpload'; // Import the new component
 import CreateTeamPopup from './CreateTeamPopup';
 import './TeamOverview.css';
 
@@ -6,7 +7,16 @@ const TeamOverview = () => {
     const [teams, setTeams] = useState([]);
     const [isPopupOpen, setIsPopupOpen] = useState(false);
     const [editIndex, setEditIndex] = useState(null); // New state for editing
-
+    
+    const handleTeamsUpload = (data) => {
+        // Process uploaded teams from CSV
+        const teamsFromCsv = data.map((row) => ({
+            name: row['Team Name'],
+            members: row['Class'] || 'N/A', // Adjust as needed
+        }));
+        setTeams(teamsFromCsv);
+    };
+    
     const handleCreateTeam = (teamName) => {
         if (editIndex !== null) {
             // Edit existing team
@@ -38,6 +48,8 @@ const TeamOverview = () => {
             <button className="create-team-btn" onClick={handleOpenPopup}>
                 {editIndex !== null ? 'Edit Team' : 'Create Teams +'}
             </button>
+            <CSVUpload onTeamsUpload={handleTeamsUpload} /> {/* Use the CSVUpload component */}
+
             <div className="teams-container">
                 {teams.map((team, index) => (
                     <div key={index} id={index} className="team-card">
