+++ CodebustersPAS-ClientApp/src/pages/Teacher/CreateTeam/styles.css
@@ -1,4 +1,3 @@
-
 body {
   font-family: Arial, sans-serif;
   background-color: #1a1a1a;
@@ -11,10 +10,10 @@ body {
   height: 100vh;
 }
 
-
 .app-container {
   text-align: center;
 }
+
 .popup-container {
   position: fixed;
   top: 0;
@@ -36,6 +35,10 @@ body {
   text-align: center;
 }
 
+.popup-content h2 {
+  margin: 0 0 20px;
+  color: white;
+}
 
 .create-team-btn {
   background-color: #4CAF50;
@@ -51,13 +54,12 @@ body {
   background-color: #45a049;
 }
 
-
 .teams-container {
   display: flex;
   justify-content: space-around;
   flex-wrap: wrap;
+  margin-top: 20px; /* Added margin for spacing */
 }
-
 
 .team-card {
   background-color: #333333;
@@ -76,31 +78,6 @@ body {
 .team-card p {
   margin: 5px 0 0;
   font-size: 14px;
-}
-
-.popup-container {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.popup-content {
-  background-color: #2b2b2b;
-  padding: 20px;
-  border-radius: 10px;
-  width: 300px;
-  text-align: center;
-}
-
-.popup-content h2 {
-  margin: 0 0 20px;
-  color: white;
 }
 
 .team-input {
@@ -164,12 +141,22 @@ body {
   color: #ccc;
 }
 
-.team-input {
-  width: 100%;
-  padding: 10px;
-  margin-bottom: 10px;
-  border-radius: 5px;
-  border: 1px solid #666666;
-  background-color: #1a1a1a;
+/* Styles for the CSV Upload Component */
+.csv-upload {
+  margin-top: 20px; /* Added margin for spacing */
+}
+
+.csv-upload input[type="file"] {
+  background-color: #333333; /* Consistent background */
   color: white;
+  padding: 10px;
+  border: 1px solid #666666;
+  border-radius: 5px;
+  cursor: pointer;
+  width: 100%;
+}
+
+.csv-upload p {
+  margin-top: 10px;
+  color: red; /* Error message color */
 }
